+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.0.0
+
+- User async/await with tokio instead of rayon. Improved performance when using --blame.
+- Upgrade crates.
+
 ## v0.8.5
 
 - --query option can be used multiple times.
@@ -14,7 +19,7 @@
 - Upgrade to Rust edition 2021.
 - Add --requested option.
 - Add the Requested (explicitly) score.
-- Simplify request query removing -author: and -reviewed-by: conditions to avoid GH bug to return an empty result.
+- Simplify request query by removing -author: and -reviewed-by: conditions to avoid GH bug to return an empty result.
 - Improve parsing of code owner team members.
 
 ## v0.8.2